Refactored MemphisClient
